Start edit, top menu and save
Can begin to edit a line by double clicking on one of it's contents

Added menu
Can save using menu. Saving writes the list to a .json file
Can quit using menu

Comiting the edit of a line is bugged, and therefore commented on this push. Hopefully will be resolved by next push.