Now working with Angular!
Thank you once again to @solidvoxel to all the help, and for once again evangelizing me to another programming language!

This commit is full of bugs, doesn't have half the funtions it's meant to have, but it's a work in progress. Hopefully, I'll get the Angular working as well (more like, working beter!) than my Python counter-part!