Removed Python Iteration from .git
For now, I'm removing the Python iteration of  improved-broccoli, as I'm focusing only the webapp version of it. Hopefully, this will make it so there is no confusion on potential users.
This, however, means there will be a major set back on the development, not on the experience I gained nor on how I'm moving on with the project, but on how much of it is viable and usable at this exact time. I strongly believe, however, that this will allow me to make a better product. Soon, it should be at least in the same condition that the python software was at. Hopefully rather soon, it will even surpass the python version.