Improvements made during GGJ19
Game Jams are where, when programmers are tired of looking at the code of the game, they look at the code of sideprojects, and then make a way t make the side project fit and be usable for the main game.

Somewhat workable version.
Saving is corrected, but trying to make it into a json key and value dictionary.
No more weird bugs with input