Save & Open Scripts
You can now save and open scripts. For now, it remembers the name of the script (it will be the same as the file's name), and the plain string of the textarea, where the script will be written.

Also, small update on the saved file names: no longer are they SB_, now they're Broccoli_